feat(ui): Migrate DataTable, children components & utils
