+++ apps/docs.blocksense.network/components/DataFeeds/DataFeedsTable.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import { FeedsConfig } from '@blocksense/config-types/data-feeds-config';
+import { DataTable } from '@blocksense/ui/DataTable';
 
 import { columns } from '@/components/DataFeeds/columns';
 import { dataFeedUrl } from '@/src/constants';
-import { DataTable } from '@/components/common/DataTable/DataTable';
 
 export const DataFeedsTable = ({ feeds }: { feeds: FeedsConfig['feeds'] }) => {
   return (
