+++ apps/docs.blocksense.network/components/DeployedContracts/proxyContractsColumns.tsx
@@ -1,10 +1,10 @@
 import { Tooltip } from '@blocksense/ui/Tooltip';
+import { ColumnDef } from '@blocksense/ui/DataTable';
+import { DataTableColumnHeader } from '@blocksense/ui/DataTable';
+import { DataTableBadge } from '@blocksense/ui/DataTable';
 
-import { DataTableColumnHeader } from '@/components/common/DataTable/DataTableColumnHeader';
 import { ContractAddress } from '@/components/sol-contracts/ContractAddress';
-import { DataTableBadge } from '@/components/common/DataTable/DataTableBadge';
 import { NetworkAddressExplorerLink } from '@/components/DeployedContracts/NetworkAddressExplorerLink';
-import { ColumnDef } from '@/components/common/DataTable/dataTableUtils';
 
 export const columns: ColumnDef[] = [
   {
