+++ libs/ts/ui/src/components/DataTable/DataTableBadge.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ReactNode } from 'react';
+import React, { ReactNode } from 'react';
 
 import { Badge } from '@blocksense/ui/Badge';
