+++ libs/ts/ui/src/components/DataTable/DataTableColumnHeader.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ReactNode } from 'react';
+import React, { ReactNode } from 'react';
 
 import { ImageWrapper } from '@blocksense/ui/ImageWrapper';
