+++ libs/ts/ui/src/components/DataTable/DataTableContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, ReactNode, useEffect, useState } from 'react';
+import React, { createContext, ReactNode, useEffect, useState } from 'react';
 
 import {
   ColumnDef,
