+++ libs/ts/ui/src/components/DataTable/DataTableFacetedFilter.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import * as React from 'react';
+import React from 'react';
 
 import { cn } from '@blocksense/ui/utils';
 import { Button } from '@blocksense/ui/Button';
