+++ libs/ts/ui/src/components/DataTable/DataTablePagination.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useContext } from 'react';
+import React, { useContext } from 'react';
 
 import { Button } from '@blocksense/ui/Button';
 import {
