+++ libs/ts/ui/src/components/DataTable/index.tsx
@@ -1,0 +1,9 @@
+export { DataTable } from './DataTable';
+export { DataTableBadge } from './DataTableBadge';
+export { DataTableColumnHeader } from './DataTableColumnHeader';
+export { DataTableProvider } from './DataTableContext';
+export { DataTableFacetedFilter } from './DataTableFacetedFilter';
+export { DataTablePagination } from './DataTablePagination';
+export { DataTableToolbar } from './DataTableToolbar';
+export type { ColumnDef } from './dataTableUtils';
+export { DataTableViewOptions } from './DataTableViewOptions';
